refactoring, cleanup, dependency update
