+++ client/src/pages/UploadBook.js
@@ -6,7 +6,7 @@ import { useGlobalContext } from '../context/OverallContext';
 import TextAreaInput from '../components/TextAreaInput';
 import ActionBtn from '../components/ActionBtn';
 import Form from '../components/Form';
-import { useBookUpload } from '../hooks/useBookUpload';
+import { useBookData } from '../hooks/useBookData';
 import Loading2 from '../components/Loading2';
 import { motion } from 'framer-motion';
 
@@ -24,7 +24,7 @@ const UploadBook = () => {
     userId,
     API_BOOKS,
   } = useGlobalContext();
-  const { bookUpload } = useBookUpload();
+  const { bookUpload } = useBookData();
 
   const textChange = (e) => {
     setNewBook({ ...newBook, [e.target.name]: e.target.value });
