+++ client/src/styles/_buttons.scss
@@ -1,0 +1,70 @@
+@mixin btn($bckgrnd-clr, $brnd-clr, $brdr-rds-3) {
+  background-color: $bckgrnd-clr;
+  border: 2px solid $brnd-clr;
+  border-radius: $brdr-rds-3;
+  color: $brnd-clr;
+  cursor: pointer;
+  transition: $trnstn;
+}
+
+@mixin hover-btn($fnt-clr) {
+  background-color: $fnt-clr;
+  border: 2px solid $fnt-clr;
+}
+
+.action-btn {
+  @include btn($bckgrnd-clr, $bckgrnd-clr, $brdr-rds);
+
+  color: $brnd-clr;
+  width: 100%;
+  margin-bottom: 0.5rem;
+  padding: 0.25rem;
+  text-transform: capitalize;
+
+  &:hover {
+    @include hover-btn($fnt-clr);
+    color: $bckgrnd-clr;
+  }
+}
+
+.filter-btn {
+  @include btn($bckgrnd-clr, $brnd-clr, $brdr-rds-3);
+
+  text-transform: uppercase;
+  padding: 0.5rem 0.5rem;
+  margin: 0.5rem 0.25rem;
+  align-self: center;
+
+  &:hover {
+    @include hover-btn($brnd-clr);
+    color: $bckgrnd-clr;
+  }
+}
+
+.signin-btn {
+  @include btn($brnd-clr, $brnd-clr, $brdr-rds);
+  color: $bckgrnd-clr;
+  padding: 0.25rem;
+  margin-top: 1rem;
+
+  &:hover {
+    @include hover-btn($fnt-clr);
+    color: $bckgrnd-clr;
+  }
+}
+
+@media (min-width: 800px) {
+  .action-btn {
+    padding: 0.5rem;
+    margin-bottom: 1rem;
+  }
+
+  .filter-btn {
+    margin: 0.5rem 0.5rem;
+  }
+
+  .signin-btn {
+    padding: 0.5rem;
+    margin-top: 1.5rem;
+  }
+}
