--- client/src/styles/_filterbtns.scss
@@ -1,19 +1,0 @@
-.filter-btn {
-  @include btn($bckgrnd-clr, $brnd-clr, $brdr-rds-3);
-
-  text-transform: uppercase;
-  padding: 0.5rem 0.5rem;
-  margin: 0.5rem 0.25rem;
-  align-self: center;
-
-  &:hover {
-    @include hover-btn($brnd-clr);
-    color: $bckgrnd-clr;
-  }
-}
-
-@media (min-width: 800px) {
-  .filter-btn {
-    margin: 0.5rem 0.5rem;
-  }
-}
