+++ client/src/styles/_search.scss
@@ -6,7 +6,7 @@
 
   .search-form {
     padding: 0.75rem;
-    border: 2px solid $bckgrnd-trnsprnt-1;
+    border: 1px solid $brnd-clr;
     border-radius: $brdr-rds-3;
     box-shadow: $mn-shdw;
   }
