--- client/src/styles/_signinbtns.scss
@@ -1,19 +1,0 @@
-.signin-btn {
-  @include btn($brnd-clr, $brnd-clr, $brdr-rds);
-  color: $bckgrnd-clr;
-  padding: 0.25rem;
-  margin-top: 1rem;
-
-  &:hover {
-    @include hover-btn($fnt-clr);
-
-    color: $bckgrnd-clr;
-  }
-}
-
-@media (min-width: 800px) {
-  .signin-btn {
-    padding: 0.5rem;
-    margin-top: 1.5rem;
-  }
-}
